update 1125
서버 - 클라 연결 확인
데이터 패키징 제대로 보내지는지 확인
현재 좌표 이상하게 전송되는 문제 발생(미해결)

+++ AirHockey_Server/term_project/term_project/Server.h
@@ -26,8 +26,8 @@ struct ClientInfo {
 
 //객체의 2차원 좌표
 struct Point2D {
-	float position_x;
-	float position_y;
+	int position_x;
+	int position_y;
 };
 
 //객체의 2차원 가속도
